feat: sicx-baln implementation
