+++ contracts/sicx/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "spoke-token"
+name = "sicx"
 version = "0.1.0"
 edition = "2021"
 
@@ -12,7 +12,8 @@ testutils = ["soroban-sdk/testutils"]
 [dependencies]
 soroban-sdk =  { workspace = true }
 soroban-token-sdk = { version = "21.6.0" }
-soroban-rlp = { path = "../soroban-rlp" }
+spoke-token = { path = "../../libs/spoke_token" }
+
 
 [dev-dependencies]
 soroban-sdk = { workspace = true, features = ["testutils"] }
