--- libs/spoke_token/src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod  setup;
-pub mod balanced_dollar_test;
